Python/mesos-unreserve/requirements.txt
certifi==2018.4.16
chardet==3.0.4
get==1.0.3
idna==2.6
post==1.0.2
public==1.0.3
python-nss==1.0.1
query-string==1.0.2
requests==2.18.4
urllib3==1.22
